6/27 17:00　更新しました

+++ src/main/java/com/example/mofmof/repository/entity/Tasks.java
@@ -21,7 +21,7 @@ public class Tasks {
     private String content;
 
     @Column
-    private short status;
+    private int status;
 
     @Column(name = "limit_date")
     private Date limitDate;
